// src/Controllers/ConversationalCloud/conversation-cloud.interfaces.ts
export interface MessageData {
  msg: {
    text: string;
  };
  quickReplies?: {
    content: string;
  };
}

export interface MessageDataRich {
  richContent: {
    content: string;
  };
}

export interface MessageRecord {
  type: MESSAGE_TYPES;
  messageData: MessageData | MessageDataRich;
  messageId: string;
  seq: number;
  dialogId: string;
  participantId: string;
  source: string;
  time: string;
  timeL: number;
  device: string;
  audience: string;
  sentBy: string;
  contextData?: {
    rawMetadata: string;
    structuredMetadata: {
      actionReason?: {
        reason: string;
      };
      botResponse?: {
        businessCases: string[];
        externalConversationId: string;
        intents: {
          confidence: string;
          confidenceScore: number;
          id: string;
        }[];
      };
    };
  };
}
// ...
export interface MessageScore {
  messageId: string;
  messageRawScore: number;
  mcs: number;
  time: string;
  timeL: number;
}
interface MessageStatus {
  messageId: string;
  seq: number;
  time: string;
  timeL: number;
  participantId: string;
  participantType: string;
  messageDeliveryStatus: string;
  dialogId: string;
}
export interface ConversationSurvey {
  surveyType: string;
  surveyStatus: string;
  dialogId: string;
  surveyData: {
    answer: string;
    answerId: string;
    answerSeq: number;
    isValidAnswer: boolean;
    question: string;
    questionFormat: string;
    questionId: string;
    questionName: string;
    questionType: string;
  }[];
}
interface CoBrowseSession {
  sessionId: string;
  startTime: string;
  startTimeL: number;
  endTime: string;
  endTimeL: number;
  endReason: string;
  duration: number;
  type: string;
  capabilities?: string[];
  agentId: string;
  interactive: boolean;
}
interface CoBrowseSessions {
  coBrowseSessionsList: CoBrowseSession[];
}

export interface CustomerInfoSde {
  serverTimeStamp: string;
  customerInfo: {
    companyBranch: string;
    customerId: string;
    customerStatus: string;
    customerType: string;
  };
}

export interface SDE {
  customerInfo?: CustomerInfoSde;
  personalInfo?: {
    originalTimeStamp: string;
    personalInfo: {
      contacts: {
        personalContact: {
          email: string;
          phone: string;
        };
      }[];
      name: string;
      surname: string;
    };
    serverTimeStamp: string;
  };
  sdeType: string;
  serverTimeStamp: string;
}

export interface SDEs {
  events: SDE[];
}